% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_through_pipe.R
\name{pprintp}
\alias{pprintp}
\title{Make pretty printout of piped code block string.}
\usage{
pprintp(ptext)
}
\arguments{
\item{ptext}{string of pipe-separated commands.}
}
\value{
pretty version of code string.
}
\description{
Make pretty printout of piped code block string.
}
