% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_through_pipe.R
\name{print_header}
\alias{print_header}
\title{Stub function which pretty-prints a call header}
\usage{
print_header(call, length)
}
\arguments{
\item{call}{A character string representing a pipe call}

\item{length}{A number indicating how long to pad the header to}
}
\description{
Stub function which pretty-prints a call header
}
