% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_through_pipe.R
\name{print_sepline}
\alias{print_sepline}
\title{Output a string of separator characters for use in print outs.}
\usage{
print_sepline(char = "-", line_length = 60)
}
\arguments{
\item{char}{Character to use.}

\item{line_length}{Number of characters in the string.}
}
\value{
None
}
\description{
Output a string of separator characters for use in print outs.
}
