% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_through_pipe.R
\name{sepline}
\alias{sepline}
\title{Make a string of separator characters to use in print outs.}
\usage{
sepline(char = "-", line_length = 60)
}
\arguments{
\item{char}{Character to use.}

\item{line_length}{Number of characters in the string.}
}
\value{
Separator string.
}
\description{
Make a string of separator characters to use in print outs.
}
