% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_through_pipe.R
\name{step_through_pipes}
\alias{step_through_pipes}
\title{Display intermediate results of blocks of piped expressions.}
\usage{
step_through_pipes(expr)
}
\arguments{
\item{expr}{Block of dplyr or other piped commands.}
}
\value{
Result of evaluating \eqn{expr}.
}
\description{
Display intermediate results of blocks of piped expressions.
}
\examples{
library(tidyverse)
step_through_pipes({
   starwars \%>\%
      mutate(bmi=mass/((height/100)^2)) \%>\%
      select(name:mass, bmi) \%>\%
      filter(bmi<22) \%>\%
      gather(feature, value, -name) \%>\%
      arrange(name, feature)
})

}
